Create modules and import them
